requirements.txt
google-generativeai
python-dotenv
flask==2.3.3
flask-cors==4.0.0